refactor: split the definitions and tests of summary objects
