+++ cobra/core/__init__.py
@@ -12,3 +12,4 @@ from cobra.core.reaction import Reaction
 from cobra.core.group import Group
 from cobra.core.solution import Solution, LegacySolution, get_solution
 from cobra.core.species import Species
+from cobra.core.summary import MetaboliteSummary, ReactionSummary, Summary
