+++ cobra/test/test_core/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import absolute_import
